Naturalization implementation for pathfinding
Right, so the pathfinding has been made to look more natural - however I'm realizing that I need to implement a more efficient algorithm for the pathfinding. There are already frame drops :/ Everything's mostly fine, it's just specifically the manner in which I get the path that needs some shoring up. Luckily, things are set so that even if I change the algorithm, I won't need to rewrite the naturalization